fix data race in split & merge code use uint32_t throughout statistic struct to keep four byte alignment make SufficientStatistics non-virtual add struct fingerprinting to debug determinism

+++ openpgl/directional/vmm/ParallaxAwareVonMisesFisherMixture.h
@@ -77,7 +77,7 @@ struct ParallaxAwareVonMisesFisherMixture
     vfloat _eMinus2Kappa[NumVectors];
     vfloat _meanCosines[NumVectors];
 
-    size_t _numComponents{maxComponents};
+    uint32_t _numComponents{maxComponents};
 
     // Parallax-aware attributes
     vfloat _distances[NumVectors];
